move menubar handling to util
